+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `SearchableList`: The item keys `experimental` and `deprecated` lead to a different rendering.
+- `Collections`, `FileFormats`, `Processes`, `SearchableList`, `ServiceTypes` and `UdfRuntimes`: Method `toggleHeading` has been introduced
 
 ### Fixed
 - `SearchableList`: Do initial search if externalSearchTerm is given